fixed_controllers

+++ controllers/contactsControllers.js
@@ -21,7 +21,8 @@ export async function getAllContacts(req, res, next) {
 export async function getOneContact(req, res, next) {
   const { id: _id } = req.params;
   try {
-    const result = await Contact.findOne({ _id });
+    // ------------------------------------------------
+    const result = await Contact.findOne({ _id, owner: req.user.id });
     if (result === null) {
       throw HttpError(404);
     }
@@ -40,7 +41,8 @@ export async function deleteContact(req, res, next) {
   const { id: _id } = req.params;
 
   try {
-    const result = await Contact.findOneAndDelete({ _id });
+    // ----------------------------------------------------
+    const result = await Contact.findOneAndDelete({ _id, owner: req.user.id });
     if (result === null) {
       throw HttpError(404);
     }
@@ -87,9 +89,14 @@ export async function updateStatusContact(req, res, next) {
   }
 
   try {
-    const result = await Contact.findOneAndUpdate({ _id }, contact, {
-      new: true,
-    });
+    // ---------------------------------------------------------
+    const result = await Contact.findOneAndUpdate(
+      { _id, owner: req.user.id },
+      contact,
+      {
+        new: true,
+      }
+    );
 
     if (result === null) {
       throw HttpError(404);
@@ -117,9 +124,14 @@ export async function updateContact(req, res, next) {
   }
 
   try {
-    const result = await Contact.findOneAndUpdate({ _id }, contact, {
-      new: true,
-    });
+    // ----------------------------------------------------------
+    const result = await Contact.findOneAndUpdate(
+      { _id, owner: req.user.id },
+      contact,
+      {
+        new: true,
+      }
+    );
 
     if (result === null) {
       throw HttpError(404);
